+++ src/components/moves/SlideShowPreviewElement.tsx
@@ -1,6 +1,9 @@
-import { TargetFile } from '@renderer/types/models/targetFile'
 import { Image } from '@chakra-ui/react'
 import { AiOutlineFile } from 'react-icons/all'
+import { useTranslation } from 'react-i18next'
+import { useSelector } from 'react-redux'
+import { TargetFile } from '@renderer/types/models/targetFile'
+import { RootState } from '@renderer/stores'
 
 interface Props {
   slideTargetFileByIndex: TargetFile | null
@@ -8,11 +11,15 @@ interface Props {
 }
 
 function MovesSlideShowPreviewElement({ slideTargetFileByIndex, assetUrl }: Props) {
+  const { t } = useTranslation()
+
+  const setting = useSelector((state: RootState) => state.moves.setting)
+
   // No file
   if (!slideTargetFileByIndex) {
     return (
       <div className="text-4xl">
-        No data
+        {t('labels.noData')}
       </div>
     )
   }
@@ -32,9 +39,18 @@ function MovesSlideShowPreviewElement({ slideTargetFileByIndex, assetUrl }: Prop
   if (slideTargetFileByIndex.type === 'video') {
     return (
       <div className="text-center max-w-full max-h-full p-2">
-        <video src={assetUrl} controls>
+        <video src={assetUrl} controls autoPlay={setting.isAutoPlay}>
           Your browser does not support the video tag.
         </video>
+      </div>
+    )
+  }
+  if (slideTargetFileByIndex.type === 'audio') {
+    return (
+      <div className="text-center max-w-full max-h-full p-2">
+        <audio src={assetUrl} controls autoPlay={setting.isAutoPlay}>
+          Your browser does not support the audio tag.
+        </audio>
       </div>
     )
   }