+++ src/router/index.tsx
@@ -7,11 +7,13 @@ import RenameLayout from '@renderer/layouts/Rename'
 import OrganizeLayout from '@renderer/layouts/Organize'
 import DeleteLayout from '@renderer/layouts/Delete'
 import { lazy } from 'react'
+import SlideLayout from '@renderer/layouts/Slide'
 
 const Moves = lazy(() => import('@renderer/pages/moves'))
 const Renames = lazy(() => import('@renderer/pages/renames'))
 const Organizes = lazy(() => import('@renderer/pages/organizes'))
 const Deletes = lazy(() => import('@renderer/pages/deletes'))
+const Slides = lazy(() => import('@renderer/pages/slides'))
 
 export default function MainRoutes() {
   return (
@@ -30,6 +32,9 @@ export default function MainRoutes() {
         </Route>
         <Route path="deletes" element={<DeleteLayout />}>
           <Route index element={<Deletes />} />
+        </Route>
+        <Route path="slides" element={<SlideLayout />}>
+          <Route index element={<Slides />} />
         </Route>
       </Route>
     </Routes>